Level 3 - 2
Calculator 클래스를 제네릭 클래스로 바꾸고 숫자들만 받을 수 있게 만들었고 이를 계산할때는 double 값으로 전환하고 계산을 할 수 있도록 구현

+++ src/main/java/calculator/AddOperator.java
@@ -4,7 +4,7 @@ package calculator;
 public class AddOperator implements Operator {
     // 인터페이스의 내용을 정의
     @Override
-    public double operate(int firstNum, int secondNum) {
+    public double operate(double firstNum, double secondNum) {
         return firstNum + secondNum;
     }
 }
