+++ src/main/java/calculator/App.java
@@ -18,14 +18,14 @@ public class App {
 
             if (choice == 1) {
                 System.out.println("첫 번째 숫자를 입력하세요.");
-                arithmetic.setFirstNum(sc.nextInt()); // Calculator 의 필드에 값저장
+                arithmetic.setFirstNum(sc.nextDouble()); // Calculator 의 필드에 값저장
 
                 System.out.println("사칙연산 기호를 입력하세요."); // +, -, *, /
                 // sc.next 가 받은 String 타입의 문자열중 charAt(i) i번 째의 값을 char 형태로 저장 ( i는 0부터 시작 )
                 arithmetic.setOperator(sc.next().charAt(0));
 
                 System.out.println("두 번째 숫자를 입력하세요.");
-                arithmetic.setSecondNum(sc.nextInt()); // Calculator 의 필드에 값저장
+                arithmetic.setSecondNum(sc.nextDouble()); // Calculator 의 필드에 값저장
 
                 // 부모 클래스 필드에 저장된 arithmetic operator 의 값을 OperatorType 에 보내서 값을 받고 그 매개변수로 메서드 작동
                 System.out.println("결과 = "+ arithmetic.calculate(OperatorType.setOperator(arithmetic.operator)));
@@ -47,7 +47,7 @@ public class App {
                 // 위의 사칙연산과 동일한 내용을 수행하도록 구현
                 // 새로운 CircleArea 리스트를 만들어서 사칙연산과 따로따로 저장되도록 구현
                 System.out.println("원의 반지름을 입력해주세요");
-                circle.setFirstNum(sc.nextInt()); // Calculator 의 필드에 값저장
+                circle.setFirstNum(sc.nextDouble()); // Calculator 의 필드에 값저장
                 System.out.println("결과 = " + circle.calculateCircleArea());
                 circle.getFirst(); // 결과값 조회
                 System.out.println("가장 먼저 저장된 결과를 삭제하시겠습니까? (remove 입력시 삭제)");
